Don't send _last_visit for identify or track
The expectation from Customer.io is that the `_last_visit` parameter
should only be sent on page requests, as it indicates when the user was
active on the site. Identify calls and event tracking should not update
this parameter.

This commit removes the calls to `visit` for identify and track, and
updates the tests to validate that identify, group, and track all only
send one request.

We still update the `_last_visit` parameter for page requests.

Fixes #6